Build an immutable memory model
The Z-machine memory model has dynamic low memory -- about 8K in the
case of Zork -- and unchanging code and strings in static / high memory.
The immutable memory model has the original state of dynamic and static
memory stored in two strings, plus an immutable map from addresses to
bytes that gives the diff between the original dynamic memory and the
new dynamic memory.   Edits to memory produce a new memory model with
the same original strings and a new map.