Make handlers more general
We know that computing the operands of an instruction can change the
interpreter, because doing so can pop the stack. Computing the result /
performing the action of the instruction can also change the
interpreter. The general-purpose instruction handlers now take a
function that both computes a result and returns the new interpreter, if
there is one.