More experiments in text rendering
I've built a maquette of the interpreter that just knows how to print
strings and ask for text, in a functional style. The interpreter knows
whether the recent prints are going to overflow the screen. If the
interpreter decides that it needs to prompt for text or for "more" then
it returns halfway through an instruction and requires the caller to
"step with input" before it continues.