Start to reorganize I/O code
So far the entire system is purely functional with the exception of the
I/O code; the interpreter synchronously waits for console input when
executing sread, and printing is the side effect of printing to the
console.

This checkin begins to move towards a more functional model. First, the
output state is gathered up into a persistent, pretty-printed list,
store in the interpreter. Any interpreter now has a snapshot of the
state of all previous outputs.  However, outputs are still made via
console output.

An unfortunate temporary problem this introduces is that echoing the
user input into the transcript echoes it to the screen, where it already
was printed.  This problem will go away when we stop using the console
for I/O.

Second, the sread code has been reorganized into two parts: the first
part simply determines the size of the buffer and produces an
interpreter in the "I need input" state. The step then ends, and the
following step is required to provide input; how the code hosting the
interpreter chooses to do so is up to it.