Implement graphical user interface
This primitive graphical user interface eliminates the reliance on
side-effecting console output and input; the state of the game is now
entirely captured in the interpreter.

Pausing long screens with --more-- is not yet implemented.

We are at present redrawing the screen on every instruction, which is
plainly too often. The interpreter should signal its caller that there
is new text to display.