A maquette of yet another screen model
I was unsatisfied with the "transcript" based output mechanism, as first
it is not sophisticated enough to handle cursor movement. Second, it is
too different from the specification; I want the concepts in the spec to
be represented more clearly by data structures. And third, because there
was too much cooperation between the interpreter and the code driving it
when dealing with [MORE] situations.

The maquette here has a type that represents a screen with a cursor. If
the screen has more data written to it than will fit, it buffers it up
and allows the display logic to scroll it one line at a time. This makes
animations of scrolling possible. Rather than calling the interpreter
back to tell it that the space was hit after [MORE], instead the calling
code is responsible for scrolling the screen should it choose to do so,
and then providing the interpreter the new screen state.

The core of the screen logic uses an immutable deque so that information
can be manipulated cheaply at both the top and the bottom of the screen