Interpreter now takes chars, not strings, as input
I did not much like this business of the caller maintaining a buffer and
looping until a whole string was obtained. Now the caller calls the
interpreter back on every keypress and the interpreter decides what the
effect of the keypress is. The caller is, however, still responsible for
printing the buffer onto the screen. (The other way to handle this would
be to have the interpreter keep two screens; the original screen without
input, and the printed screen with.)