Scrolling now done by interpreter, not caller
Similarly to the previous checkin, I was dissatisfied with how the
interpreter hands back a screen that needs scrolling, the caller scrolls
it, and then hands back the scrolled screen to the interpreter.

Now the interpreter hands back a screen that needs scrolling, and then
each "step" gives back the scrolled screen until no more scrolling is
required.

The interpreter also tracks whether "more" is needed, but allows the
caller to be the one to display the "more".

There are a great many changes in this diff because I'm trying out the
atom editor and it strips extraneous spaces.