Implement restore
This necessitated a number of changes:

* The stack frame was storing the address of the call, not the address
of the instruction after the call. This was so that the store logic
could be looked up in the calling instruction upon return.  Apparently
this is not how most interpreters do it. Quetzal files expect that the
frame record will contain the address *after* the call instruction,
where execution will actually resume, and that the frame must itself
contain the location for the store. That way the return does not have to
know what the call instruction was.  I've implemented this change.

* The stack frame must also store the number of arguments provided;
we'll need this information when we implement saving.

* Restore is now implemented but with no error handling whatsoever. It
will need to be refactored somewhat. Exception handling is likely the
most convenient mechanism.

* Numerous minor improvements to crash diagnosis and debugging dumpers