Fix nasty bug in local / stack storage
When writing a word to story memory we have no choice but to truncate it
to 16 bits. But locals and stack memory are full on ints, which means
they can be negative numbers, results of overflowed multiplications, and
so on.  We should truncate back to unsigned 16 bit integers before
storing to the frame; they can be turned back into signed quantities if
necessary later.

Also, fortified the memory subsystem so that it detects out-of-bounds
accesses more aggressively.