Large refactoring of the screen model
What was a "screen" is now a "window".  A "screen" is composed of two
"windows" which are non-overlapping rectangles of text with independent
cursor, word wrapping semantics, and so on. Scrolls on the lower window
do not affect the upper window.

Window could stand to be moved to a module of its own.