Refactor window and button code to use wrapper types
We now have wrapper types representing pixel and character widths,
heights and coordinates.

Also added a cursor type, which could use some more work.

The code is still rough and needs much refactoring to make it pretty,
but at least now we know that we're not mixing characters with pixels or
widths with coordinates.